ODK-120: Changed the way selected form ids are passed for sync
These are the same changes as on the repos in which survey and services where separated from each other.

Now form instances ids are passed in a list which is the value for the map, the main form id is the key.
Now only the tables which contain selected form instances are sync with the server.

Conflicts:
	survey_app/src/main/res/values/strings.xml