ODK-116: Removed duplicate strings after merging survey and services and unused files
This is a major clean-up of the string.xml files. The full list of used strings is defined in the default strings.xml in the res/values directory.
The rest of the xml files contain only a part of the main list of strings.

It is related to the ODK-39 issue and dealing with the translation of the app. This is handled by Android after changing the system language. All strings that have their equivalents would change. Those which do not have them would stay in English (from res/values/strings.xml file).

Removed unused strings and files from the project.

Conflicts:
	survey_app/src/main/res/values/colors.xml
	survey_app/src/main/res/values/strings.xml