ODK-150: Added navigation drawer & toolbar
For now, onNavigationItemSelected isn't fully implemented as we do not have those fragments to refer to.
After implementing them, we will have to make sure that the Settings get unchecked when quitting it.
Also, I've touched some design_navigation dimensions, which should be kept unchanged as those are defined by material design principles, and may be changed or removed in future, so watch out.
Some changes to build.gradle were also required.