ODK-77: Made subforms which are related to synced forms, sync as well
Enabled syncing subforms alongside with the selected forms to sync. The main problem was to create relations between forms and subforms as this information was kept only in the form definition. This is achieved by having two local tables: _main_form and form_subform_pairs. Whenever a new main form is opened/edited its basic ionformation (UUID, tableId) is written in the _main_form table.
During the process of filling/editing form, when a subform is saved then its relation with the main form is saved in the _form_subform_pairs table.
When user pick a form (which contains subforms) to sync, its UUID is added to a list. Before the sync process, _form_subform_pairs table is checked whether a main form which is going to be synced has any relations to a subform. If yes, the subform UUID is added to the sync list. If there are many relations to the same main form then all subforms UUIDs are added to the list.
After succesful sync, rows in _form_subform_pairs table which represent relations of the synced forms and subforms are removed from the table.

ODK-114: Added receiving main form data form the WebView side

Next step: save it to the local form-subform table

ODK-114: Added saving main form data and form-subform pairs in database

Added setting main form data at the beginning of completing/editing a main form.
If the main form contains a subform, on save the relation is saved in the form-subform table.

Made ServiceConnectionWrapper a public class

This service connection wrapper helps creating connection with the local database on the Android device.

ODK-114: Added removing outdated relations in the form-subform table

Relations (rows) are removed when one side of the relation is removed (form or a subform). If a form which gets deleted was connected to many subforms, all relations are deleted. This way subforms would not be included in the sync process with the server.
Additionaly, changed the name of the column in form-subform table from "subform_id" to "subform_uuid" as UUID are stored within it.

ODK-114: Updated systemzip and configzip files

ODK-114: Removed creation of the form-subform table from the AppSynchronizer

This functionality is done within the OdkSurvey

ODK-115: Added subforms related to forms so they can be synced alongside

Added subforms related to forms which were selected to sync, so they will be synced as well. Their UUIDs are taken from the form-subform local table and added to the list which is passed to sync process.
After sucessful sync, these relations between forms and subforms are removed from the form-subform local table.