ODK-220: Fixed the errors while downloading/sending files to server
Fixed the "Internal error" during the download forms process. This required blocking the attempt to synchronize the user created rows from forms tables as no form is choosen to be sync when we download form definitions.
Fixed the "Internal error" which occured during the sync form process. The creation of the _form_subform_pairs has been moved to a correct place making sure that we do not try to check contents of the non-exisiting table.