ODK-164: Added connection with the server via SyncActivity
It is possible to download form definitions by selecting the Check For Updates button and upload choosen completed form after selecting the cloud image on the In Progress Surveys list. However, both actions are launching the SyncActivity from the old Services app.
TODO: remove the SyncActivity being launched in the middle of the process and call directly the sync service methods.