ODK-163: Reworked the download of the form definitions
OdkSyncService is called directly from the MainMenuActivity so the download of the form definitions is done in background. User is informed by the notifications on the notification bar and also a final result is presented when the sync is finished.
TODO:
Make sure that after the sync user is redirected to the previous page. BackStack is sometimes causing a little mess in screen history.