ODK-257: Fixed the need to restart the app to view the downloaded forms
Now the initialization dialog is launched only after the end of work the sync process in the sync service. The app keeps on picking the service about the sync status and when it gets response other than NONE or SYNCING it terminates the sync service and starts processing the downloaded data.
Depending on what type of response it gets a different Toast message is displayed to the user along with the notification on the notification bar.
Along with fixing the download process, the upload was also fixed. Now after successful upload of a filled form a list of In Progress Instances refreshes itself with results in the synced form being removed from and it and moved to the Submitted list.

TODO: Fix the backstack problem as the INITILIZATION_DIALOG stays on it and it shouldn't. So only the first sync the process completes without any strange behavior. This is valid only for the download process as the upload handles correctly.