ODK-269: Added saving changes made within a form when changing form screens due to pressing next/back buttons
Added saving changes as INCOMPLETE which were made within the form before the screen changes when the next/back button is pressed. This updates the stored values within the database and prevents from situation in which there are many unsaved rows representing the same form (they have the same uuid) in different points in time.